requirements.txt
# Core dependencies
Django>=3.2,<5.0
djangorestframework==3.15.2
django-filter==24.3
drf-yasg==1.21.8
djangorestframework-jsonapi==7.1.0
mysqlclient==2.2.5
six==1.16.0
gunicorn==20.1.0

# Compatibility package
six==1.16.0

# Testing dependencies
pytest==8.3.3
pytest-django==4.9.0
coverage==7.6.4

